Restructure project to work with routes
